Add file textproto/reader_email.go
